Refactor: Moved Reports and Reciepts to Main Level.
