+++ semesterstat/api/bulk.py
@@ -3,7 +3,6 @@ from typing import List
 from fastapi import APIRouter, Depends, status
 from sqlalchemy.orm import Session
 
-from ..common.reports import DepartmentReport, ScoreReport, StudentReport, SubjectReport
 from ..crud.bulk import (
     put_department_bulk,
     put_score_bulk,
@@ -11,6 +10,7 @@ from ..crud.bulk import (
     put_subject_bulk,
 )
 from ..database import get_db
+from ..reports import DepartmentReport, ScoreReport, StudentReport, SubjectReport
 
 bulk = APIRouter()
