+++ semesterstat/generator.py
