+++ semesterstat/reports.py
@@ -2,12 +2,12 @@ from typing import List, Optional
 
 from pydantic import BaseModel, validator
 
-from .subject_extractor import (
+from .common.subject_extractor import (
     get_subject_dept,
     get_subject_scheme,
     get_subject_semester,
 )
-from .usn_extractor import get_usn_batch, get_usn_dept
+from .common.usn_extractor import get_usn_batch, get_usn_dept
 
 # https://pydantic-docs.helpmanual.io/usage/validators/#validate-always
