+++ tests/docgen/conftest.py
@@ -3,8 +3,8 @@ import json
 import pytest
 from sqlalchemy.orm import Session
 
-from semesterstat.common.reports import ScoreReport, StudentReport, SubjectReport
 from semesterstat.database import Score, Student, Subject
+from semesterstat.reports import ScoreReport, StudentReport, SubjectReport
 
 
 @pytest.fixture(scope="package", autouse=True)
